pyproject.toml
[project]
name = "treding-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "cryptography>=46.0.5",
    "dotenv>=0.9.9",
    "python-binance>=1.0.35",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "typer>=0.24.1",
]
